Add missing failure method to BaseFlow
- This seemed a better option than making BaseFlow an abstract class raising
NoMethodError, since it looks like the logic would be the same anyhow.
