+++ app/services/flow/base_flow.rb
@@ -1,5 +1,7 @@
 module Flow
   class BaseFlow
+    include Failure
+
     attr_accessor :flow_session
     attr_reader :steps, :actions, :current_user, :current_sp, :params, :request, :json,
                 :http_status, :controller
